+++ test/screenshot/webpack.config.js
@@ -1,15 +1,15 @@
 module.exports = [
-  ...require('./button/webpack.config.js'),
-  ...require('./card/webpack.config.js'),
-  ...require('./chips/webpack.config.js'),
-  ...require('./fab/webpack.config.js'),
-  ...require('./floating-label/webpack.config.js'),
-  ...require('./line-ripple/webpack.config.js'),
-  ...require('./material-icon/webpack.config.js'),
-  ...require('./notched-outline/webpack.config.js'),
+  // ...require('./button/webpack.config.js'),
+  // ...require('./card/webpack.config.js'),
+  // ...require('./chips/webpack.config.js'),
+  // ...require('./fab/webpack.config.js'),
+  // ...require('./floating-label/webpack.config.js'),
+  // ...require('./line-ripple/webpack.config.js'),
+  // ...require('./material-icon/webpack.config.js'),
+  // ...require('./notched-outline/webpack.config.js'),
   ...require('./select/webpack.config.js'),
-  ...require('./text-field/webpack.config.js'),
-  ...require('./text-field/helper-text/webpack.config.js'),
-  ...require('./text-field/icon/webpack.config.js'),
-  ...require('./top-app-bar/webpack.config.js'),
+  // ...require('./text-field/webpack.config.js'),
+  // ...require('./text-field/helper-text/webpack.config.js'),
+  // ...require('./text-field/icon/webpack.config.js'),
+  // ...require('./top-app-bar/webpack.config.js'),
 ];
