--- test/unit/select/Select.test.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import td from 'testdouble';
-import {assert} from 'chai';
-import {mount, shallow} from 'enzyme';
-import Select from '../../../packages/select/Select';
-
-suite('Select Parent Container');
-
-test('has mdc-select class', () => {
-  const wrapper = shallow(<Select label='my label' />);
-  assert.isTrue(wrapper.hasClass('mdc-select'));
-});
-
-test('classNames adds classes', () => {
-  const wrapper = shallow(<Select
-    label='my label'
-    className='test-class-name'
-  />);
-  assert.isTrue(wrapper.hasClass('test-class-name'));
-});
